Fix - Database restore fails with `is empty`
Short Summary
When the container is built from Docker Hub, the file `WideWorldImporters-Full.bak` is
not copied into the container. This is because docker hub, doesn't support GIT LFS.

Error Message: `The volume on device '/var/opt/sqlserver/WideWorldImporters-Full.bak' is empty.`

Why is it necessary? (Bug fix, feature, improvements?)
- Database was not restoring

How does the change address the issue?
- Using ADD which can access files from URL's

Include a link to the ticket, if any.
- https://github.com/docker/hub-feedback/issues/500